enabled guild_member intent to properly manage members, minor improvements
